pyproject.toml
[tool.poetry]
name = "wallet_transactions"
version = "0.1.0"
description = ""
authors = ["Kirill Ilichev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-rest-framework = "^0.1.0"
django-filter = "^24.2"
djangorestframework-jsonapi = "^7.0.0"
django-environ = "^0.11.2"
pymysql = "^1.1.1"
importlib-metadata = "^4.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.0"
black = "^24.4.2"

[tool.black]
line-length = 92
target-version = ['py311']
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | env
  | __pycache__
  | migrations
  | dist
  | build
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
